Updated Lookup Plugin for UDDI Ansible (#35)
* Modified Lookup Plugin for UDDI

* Fixed LINT errors

* Added function to return_base_url

* Added support for Environment variables config

* fixed linter errors

* Fixed linter errors

* Reverted back tests

* Reverted unnecessary files

* Updated Version for Lookup Plugin

* Addressed Review comments

* Modified Return

* Corrected Examples

* Addressed Review comments

---------
